Update types of legacy tracker model to match to domain one (#245)
* `score` to Double

* `tracker_id` to Long

* `last_chapter_read` to Double

* `total_chapters` to Long

* `status` to Long

Additional changes for Komikku/TachiyomiSY version

* `score` to Double

* `tracker_id` to Long

* `last_chapter_read` to Double

* `total_chapters` to Long

* `status` to Long

+++ app/src/main/java/eu/kanade/domain/track/interactor/TrackChapter.kt
@@ -34,7 +34,7 @@ class TrackChapter(
                     service == null ||
                     !service.isLoggedIn ||
                     chapterNumber <= track.lastChapterRead /* SY --> */ ||
-                    (service is MdList && track.status == FollowStatus.UNFOLLOWED.int.toLong())/* SY <-- */
+                    (service is MdList && track.status == FollowStatus.UNFOLLOWED.long.toLong())/* SY <-- */
                 ) {
                     return@mapNotNull null
                 }
